Create global component for modals (#20)
